requirements.txt
#django==1.8.13
pillow==3.4.2
django-bootstrap3
django-allauth==0.31.0
WeasyPrint
